casts

+++ app/RiderLocation.php
@@ -12,8 +12,8 @@ class RiderLocation extends Model
     protected $fillable = ['latitude','longitude', 'rider_id', 'trip_id'];
 
     protected $casts = [
-        'latitude' => 'integer',
-        'longitude' =>'integer'
+        'latitude' => 'float',
+        'longitude' =>'float'
     ];
 
    	public function user(){
